added firewall address object management
you're now able to generate address object names from various
informations (you may build your own pattern from the given informations
in the firewall zone management settings section). fw object (firewall
address object) field in the ip address table is mandatory but can be
disabled by filter ip field switches. you may also disable the automatic
generation of fw objects during the creation of a ip address object and
generate it manually for just single addresses or for the whole subnet
instead. fw objects are persistent even if the information,  which are
used to by the pattern to create the name of the object, changes. just
regenerate the fw object name or leave it as it is (if there is the need
to).