fix draw_rack when adding new device and deviceId is empty
this causes a broken image in the new device dialog when selecting a rack and display it without the check for empty value:
app/tools/racks/draw_rack.php?rackId=2&deviceId=

And as the manual for php says it should be enough to only use empty():
No warning is generated if the variable does not exist. That means empty() is essentially the concise equivalent to !isset($var) || $var == false.