fix: Improve LDAP authentication process
This is an attempt and a suggestion to improve the user authentication
with LDAP.

We have a baseDN e.g. "o=ourcompany.com" and several trees, e.g.

    "ou=admins,ou=stuff,o=ourcompany.com"
    "ou=operators,ou=noc,o=ourcompany.com"

All users within these trees have permission to login to phpipam, the
use their username (say "john") and their password. As we have two trees
we cannot use just the "accountSuffix" setting as they are different.

This patch supposes an authentication by binding as "adminUsername" or
anonymously as configured, then lookup the given UID below "baseDn". It
is forced to be unique here. The actual authentication is made with the
DN of the UID and the supplied password.