Add Alpine Dockerfile
This adds a docker file that can be used to make deploying the application
easier. It does not focus on trying to be an all in one solution as you
still need to setup mysql. The full image is only 55.5MB and makes a small
change to the config file that lets you use environment variables to set
config values all namespaced with IPAM_* for safty.