Add more notes from a night of pain
rebased and ran into a bunch of issues since submodules are now required and
no errors are visible in the UI when they fail to load in. For whatever reason
on initial install the password will not set I have tried 3 different times
something is going on that I will try to debug but i can see after i reset the password
an error about php mailer is puked out so likely something to do with that (it's about
an email not being valid).

Some work also needs to be done around the database install process to better support
docker out of the box. The automated install is super nice but only allowing access to
localhost through this method means you have to log into mysql and run commands. Adding
an additional checkbox to the process like "allow remote access" or otherwise would fix
this up. I could see some security concerns around this since it is opening up mysql to
the world however if you do not have your firewall setup or are not running your db in a
container where the port is hidden from the outside world.