api: Use array_key_exists instead of isset for address PATCH merge
isset(array[key]) returns false when array[key] = null, even if the key
exists, so it seems. It looks like we are looking for the presence here
(or lack thereof) of a key in the update data, so this should suffice.

This was causing a bug where setting a custom field to null to unset it
was not functioning - a regression in the addresses controller since
1.2.