Fix ip_addr and dns_name mappings
The frontend states that the fields should be named "Hostname" and "IP Address" in excel, but then fails to match them during import because they're called "ip_addr" and "dns_name" internally (and in the DB). 

This change allows using either the database format or the one displayed in the frontend in excel sheets for import.