Enbale L2 Domain import.
Fix all templates so that header fields in the template so that all
the template fields provided are automatically matched.

This needed to change the java script, as this matching is done client side,
removing a string replacement of HOSTNAME to DNS_NAME.  Doesn't seem to
break any other importing, but not widely tested.