footer.php: remove non working usage of l10n macro
With dd2aeb438 a option for showing a custom footer message was
introduced. The usage of the _() macro isn't and can't work that way.
And there is no need to use localization here, the user will use a
"customized" message, means he will use already use the preferred
language!
And even if the message is saved within the database, the user will also
here use his preferred language.