Enhancement: edit device front/back rack position labels
In conjunction with changes made in edit-rack-devices.php adding a device to the rack, the start position had two option groups, a front, and a back group.  It was not intuitive that there were two option groups in the dropdown list. When a start position was selected it showed '5' and you did not know was it a front or a back 5.  Added 'Front:" and 'Back:' to the select option.