Feature: remove_offline_addresses.php only autodiscovery
Added an option to clean only autodiscovered and offline addresses

This is based on description 

This is useful when you want to clean "accidentally" autodiscovered IPs that are now reported offline without affecting other offline entries.

Not sure though if it's the best way to do it. AFAIK checking the description is the only way to identify the entry as autodiscovered (when the description never got changed to something else).